requirements.txt
kafka-python==1.4.5
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
